fix listeners

+++ packages/twenty-server/src/modules/calendar/blocklist-manager/jobs/blocklist-item-delete-calendar-events.job.ts
@@ -95,6 +95,10 @@ export class BlocklistItemDeleteCalendarEventsJob {
         },
       });
 
+    if (calendarEventsAssociationsToDelete.length === 0) {
+      return;
+    }
+
     await calendarChannelEventAssociationRepository.delete(
       calendarEventsAssociationsToDelete.map(({ id }) => id),
     );
